Change id type int to long

+++ SolsticeCodeChallengeWebAPI/Controllers/ValuesController.cs
@@ -19,7 +19,7 @@ namespace SolsticeCodeChallengeWebAPI.Controllers
 
         // GET api/values/5
         [HttpGet("{id}")]
-        public ActionResult<string> Get(int id)
+        public ActionResult<string> Get(long id)
         {
             return "value";
         }
@@ -32,13 +32,13 @@ namespace SolsticeCodeChallengeWebAPI.Controllers
 
         // PUT api/values/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] string value)
+        public void Put(long id, [FromBody] string value)
         {
         }
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public void Delete(long id)
         {
         }
     }
